==> R/02.2_trends-gams-plots.R <==
# load model files --------------------------------------------------------

load(file = here("output", "models", "fm.mod.RData")) 
load(file = here("output", "models", "pc.mod.RData")) 
load(file = here("output", "models", "pi.mod.RData"))
load(file = here("output", "models", "ss.mod.RData")) 


# overall trend plots -----------------------------------------------------

# modifications of `wqtrends::show_prdseries_mod()` fxn

show_prdseries_mod <-  function (mod, alpha = 0.7, base_size = 11, xlim = NULL, 
                                 ylim = NULL) 
{
  prds <- anlz_prd(mod)
  trans <- unique(prds$trans)
  tobacktrans <- mod$model %>% dplyr::mutate(trans = mod$trans)
  moddat <- anlz_backtrans(tobacktrans) %>% dplyr::mutate(date = lubridate::date_decimal(cont_year), 
                                                          date = as.Date(date))
  p <- ggplot2::ggplot(prds, ggplot2::aes(x = date)) + 
    ggplot2::geom_point(data = moddat, 
                        ggplot2::aes(y = value), shape = 1, color = "black") + 
    ggplot2::geom_line(ggplot2::aes(y = value), 
                       linewidth = 0.75, alpha = alpha, colour = "darkgreen") + 
    ggplot2::theme_bw(base_family = "sans", 
                      base_size = base_size) + 
    ggplot2::theme(legend.position = "top", 
                   legend.title = ggplot2::element_blank(), 
                   axis.title.x = ggplot2::element_blank(),
                   axis.text = element_text(color = "black"),
                   plot.margin = unit(c(1,1,1,1), "pt")) +
    ggplot2::scale_x_date(date_minor_breaks = "years") +
    ggplot2::labs(y = "Chl-a (\U00B5g/L)") + 
    ggplot2::coord_cartesian(xlim = xlim, ylim = ylim)
  if (trans != "ident") 
    p <- p + ggplot2::scale_y_log10()
  return(p)
}

# modification of `wqtrends::show_trndseason()` fxn

show_trndseason_mod <- function (mod, metfun = mean, doystr = 1, doyend = 364, 
                                 type = c("log10", "approx"), justify = c("left", "right", "center"), win = 5, 
          ylab, nsim = 10000, useave = FALSE, base_size = 11, xlim = NULL, 
          ylim = NULL, ...) 
{
  justify <- match.arg(justify)
  type <- match.arg(type)
  trndseason <- anlz_trndseason(mod = mod, metfun = metfun, 
                                doystr = doystr, doyend = doyend, justify = justify, 
                                win = win, nsim = nsim, useave = useave, ...)
  dts <- as.Date(c(doystr, doyend), origin = as.Date("2000-12-31"))
  strt <- paste(lubridate::month(dts[1], label = T, abbr = T), 
                lubridate::day(dts[1]))
  ends <- paste(lubridate::month(dts[2], label = T, abbr = T), 
                lubridate::day(dts[2]))
  subttl <- paste0("Estimates based on ", justify, " window of ", 
                   win, " years")
  yrrng <- range(trndseason$yr, na.rm = T)
  toplo <- trndseason %>% dplyr::mutate(pval = dplyr::case_when(pval < 
                                                                  0.05 ~ "p < 0.05", T ~ "ns"), pval = factor(pval, levels = c("ns", 
                                                                                                                               "p < 0.05"))) %>% na.omit()
  if (type == "log10" & mod$trans == "log10") {
    ttl <- paste0("Annual log-slopes (+/- 95%) for seasonal trends: ", 
                  strt, "-", ends)
    p <- ggplot2::ggplot(data = toplo, ggplot2::aes(x = yr, 
                                                    y = yrcoef, fill = pval)) + ggplot2::geom_hline(yintercept = 0) + 
      ggplot2::geom_errorbar(ggplot2::aes(ymin = yrcoef_lwr, 
                                          ymax = yrcoef_upr, color = pval), width = 0) + 
      ggplot2::scale_color_manual(values = c("black", "tomato1"), 
                                  drop = FALSE)
  }
  if (type == "approx" & mod$trans == "log10") {
    ttl <- paste0("Annual slopes (approximate) for seasonal trends: ", 
                  strt, "-", ends)
    p <- ggplot2::ggplot(data = toplo, ggplot2::aes(x = yr, 
                                                    y = appr_yrcoef, fill = pval)) + ggplot2::geom_hline(yintercept = 0) + 
      ggplot2::labs(title = ttl, subtitle = subttl, y = ylab)
  }
  if (mod$trans == "ident") {
    ttl <- paste0("Annual slopes (+/- 95%) for seasonal trends: ", 
                  strt, "-", ends)
    p <- ggplot2::ggplot(data = toplo, ggplot2::aes(x = yr, 
                                                    y = yrcoef, fill = pval)) + ggplot2::geom_hline(yintercept = 0) + 
      ggplot2::geom_errorbar(ggplot2::aes(ymin = yrcoef_lwr, 
                                          ymax = yrcoef_upr, color = pval), width = 0) + 
      ggplot2::scale_color_manual(values = c("black", "tomato1"), 
                                  drop = FALSE)
  }
  p <- p + ggplot2::geom_point(shape = 21, size = 3) + ggplot2::scale_fill_manual(values = c("white", 
                                                                                             "tomato1"), drop = FALSE) + ggplot2::scale_x_continuous(limits = yrrng) + 
    ggplot2::theme_bw(base_family = "sans", 
                      base_size = base_size) + 
    ggplot2::theme(axis.title.x = ggplot2::element_blank(), 
                   legend.position = "bottom", 
                   legend.title = ggplot2::element_blank(),
                   axis.text = ggplot2::element_text(color = "black", size = 12)) + 
    ggplot2::labs(title = ttl, subtitle = subttl, y = ylab) + 
    ggplot2::coord_cartesian(xlim = xlim, ylim = ylim)
  return(p)
}

# modification of `wqtrends::show_metseason()` fxn

show_metseason_mod <- function (mod, metfun = mean, doystr = 1, doyend = 364, yrstr = 2000, 
          yrend = 2019, yromit = NULL, ylab, nsim = 10000, useave = FALSE, 
          base_size = 11, xlim = NULL, ylim = NULL, ...) 
{
  chk <- identical(deparse(metfun), deparse(mean))
  if (!chk & useave) 
    stop("Specify metfun = mean if useave = T")
  if (useave) 
    metseason <- anlz_avgseason(mod, doystr = doystr, doyend = doyend)
  if (!useave) 
    metseason <- anlz_metseason(mod, metfun, doystr = doystr, 
                                doyend = doyend, nsim = nsim, ...)
  if (!is.null(yromit)) 
    metseason <- metseason %>% dplyr::filter(!yr %in% yromit)
  trans <- mod$trans
  dts <- as.Date(c(doystr, doyend), origin = as.Date("2000-12-31"))
  strt <- paste(lubridate::month(dts[1], label = T, abbr = T), 
                lubridate::day(dts[1]))
  ends <- paste(lubridate::month(dts[2], label = T, abbr = T), 
                lubridate::day(dts[2]))
  func <- as.character(substitute(metfun))
  ttl <- paste0("Est. ", func, " with 95% confidence intervals: ", 
                strt, "-", ends)
  subttl <- NULL
  toplo1 <- metseason
  p <- ggplot2::ggplot(data = toplo1, ggplot2::aes(x = yr, 
                                                   y = bt_met)) + 
    ggplot2::geom_point(colour = "black") + 
    ggplot2::geom_errorbar(ggplot2::aes(ymin = bt_lwr, ymax = bt_upr), 
                           colour = "black") + 
    ggplot2::theme_bw(base_family = "sans", 
                      base_size = base_size) + 
    ggplot2::theme(axis.title.x = ggplot2::element_blank(),
                   axis.text = ggplot2::element_text(color = "black"),
                   plot.margin = unit(c(1,1,1,1), "pt"))
  if (!any(is.null(yrstr) | is.null(yrend))) {
    mixmet <- anlz_mixmeta(metseason, yrstr = yrstr, yrend = yrend, 
                           yromit = yromit)
    toplo2 <- data.frame(yr = seq(yrstr, yrend, length = 50)) %>% 
      dplyr::mutate(met = predict(mixmet, newdata = data.frame(yr = yr)), 
                    se = predict(mixmet, newdata = data.frame(yr = yr), 
                                 se = T)[, 2], bt_lwr = met - 1.96 * se, bt_upr = met + 
                      1.96 * se, bt_met = met)
    pval <- coefficients(summary(mixmet)) %>% data.frame %>% 
      .[2, 4] %>% anlz_pvalformat()
    if (mod$trans == "log10") {
      dispersion <- summary(mod)$dispersion
      toplo2 <- data.frame(yr = seq(yrstr, yrend, length = 50)) %>% 
        dplyr::mutate(met = predict(mixmet, newdata = data.frame(yr = yr)), 
                      se = predict(mixmet, newdata = data.frame(yr = yr), 
                                   se = T)[, 2], bt_lwr = 10^((met - 1.96 * 
                                                                 se) + log(10) * dispersion/2), bt_upr = 10^((met + 
                                                                                                                1.96 * se) + log(10) * dispersion/2), bt_met = 10^(met + 
                                                                                                                                                                     log(10) * dispersion/2))
      slope <- lm(bt_met ~ yr, toplo2) %>% summary %>% 
        coefficients %>% .[2, 1]
      slope <- round(slope, 2)
      logslope <- summary(mixmet)$coefficients[2, c(1, 
                                                    5, 6)]
      logslope <- round(logslope, 2)
      logslope <- paste0(logslope[1], " (", logslope[2], 
                         ", ", logslope[3], ")")
      # subttl <- paste0("Trend from ", yrstr, " to ", yrend, 
      #                  ": approximate slope ", slope, ", log-slope ", 
      #                  logslope, ", ", pval)
      # subttl <- paste0("Approximate slope ", slope, ", log-slope ", 
      #                  logslope, ", ", pval)
      subttl <- NULL
    }
    if (mod$trans == "ident") {
      slope <- summary(mixmet)$coefficients[2, c(1, 5, 
                                                 6)]
      slope <- round(slope, 2)
      slope <- paste0(slope[1], " (", slope[2], ", ", slope[3], 
                      ")")
      subttl <- paste0("Trend from ", yrstr, " to ", yrend, 
                       ": slope ", slope, ", ", pval)
    }
    p <- p + ggplot2::geom_ribbon(data = toplo2, 
                                  ggplot2::aes(ymin = bt_lwr, 
                                               ymax = bt_upr), fill = "tomato", alpha = 0.4) + 
      ggplot2::geom_line(data = toplo2, color = "tomato")
  }
  p <- p + ggplot2::labs(title = ttl, subtitle = subttl, y = ylab) + 
    ggplot2::coord_cartesian(xlim = xlim, ylim = ylim)
  return(p)
}

# create series plots -----------------------------------------------------

plot1 <- function(mod, site, save) {
  ylab <- "Chl-a (\U00B5g/L)"
  
  site <- paste(site)
  filename <- paste0(site,".plots.png")
  
  a <- show_prdseries_mod(mod)
  b <- show_metseason_mod(mod, doystr = 1, doyend = 365, yrstr = 2003, yrend = 2022, ylab = ylab) + labs(title = "") 
  c <- show_trndseason_mod(mod, doystr = 1, doyend = 365, justify = 'center', win = 3, 
                       ylab = 'Log10 chl-a change/yr, average') + labs(title = "", subtitle = "")
  
  d <- a / b / c + plot_annotation(tag_levels = "A")
  
  if (save == TRUE) {
    ggsave(d, filename = here('output', 'figures', 'trends', filename))
  }
  else {
    print(d)
  }
  
}

plot1(mod = pi.mod, site = "PI", save = F)
plot1(mod = ss.mod, site = "SS", save = F)
plot1(mod = fm.mod, site = "FM", save = F)
plot1(mod = pc.mod, site = "PC", save = F)


# figures -----------------------------------------------------------------

# marine stations
# 
ss.gam <- show_prdseries_mod(ss.mod) + labs(title = "A")
ss.trd <- show_metseason_mod(ss.mod, doystr = 1, doyend = 365, 
                             yrstr = 2003, yrend = 2022, 
                             ylab = ylab) + labs(title = "C")

fm.gam <- show_prdseries_mod(fm.mod) + labs(title = "B", y = "")
fm.trd <- show_metseason_mod(fm.mod, doystr = 1, doyend = 365, 
                             yrstr = 2003, yrend = 2022, 
                             ylab = ylab) + labs(title = "D",
                                                 y = "")
fig2 <- 
  (ss.gam + fm.gam) / 
  (ss.trd + fm.trd)

ggsave(fig2, filename = here('output', 'figures', 'finals', 'figure2.png'),
       dpi = 600, units = "in",
       width = 6.5, height = 5)

# freshwater stations
# 
ylab <- "Chl-a (\U00B5g/L)"

pi.gam <- show_prdseries_mod(pi.mod) + labs(title = "A")
pi.trd <- show_metseason_mod(pi.mod, doystr = 1, doyend = 365, 
                             yrstr = 2003, yrend = 2022, 
                             ylab = ylab) + labs(title = "C")

pc.gam <- show_prdseries_mod(pc.mod) + labs(title = "B", y = "")
pc.trd <- show_metseason_mod(pc.mod, doystr = 1, doyend = 365, 
                             yrstr = 2003, yrend = 2022, 
                             ylab = ylab) + labs(title = "D",
                                                 y = "")
fig3 <- 
  (pi.gam + pc.gam) / 
  (pi.trd + pc.trd)

ggsave(fig3, filename = here('output', 'figures', 'finals', 'figure3.png'),
       dpi = 600, units = "in",
       width = 6.5, height = 5)
